chore: update references to replace previous testSetup
