+++ test/specs/anInitialSuite.e2e.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { step } from 'mocha-steps';
-import { refactoredTestSetup } from '../refactoredTestSetup.ts';
+import { TestSetup } from '../testSetup.ts';
 import * as utilities from '../utilities/index.ts';
 
 /*
@@ -23,7 +23,7 @@ suite does run, it needs to run first.
 */
 
 describe('An Initial Suite', async () => {
-  const testSetup = new refactoredTestSetup();
+  const testSetup = new TestSetup();
 
   step('Install extensions', async () => {
     await utilities.installExtensions();
