+++ test/specs/auraLsp.e2e.ts
@@ -5,14 +5,14 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { step } from 'mocha-steps';
-import { refactoredTestSetup } from '../refactoredTestSetup.ts';
+import { TestSetup } from '../testSetup.ts';
 import * as utilities from '../utilities/index.ts';
 
 import { Key } from 'webdriverio';
 const CMD_KEY = process.platform === 'darwin' ? Key.Command : Key.Control;
 
 describe('Aura LSP', async () => {
-  const testSetup = new refactoredTestSetup();
+  const testSetup = new TestSetup();
 
   const testReqConfig: utilities.TestReqConfig = {
     projectConfig: {
