+++ test/specs/createProjectTest.e2e.ts
@@ -1,9 +1,9 @@
 import { step } from 'mocha-steps';
-import { refactoredTestSetup } from '../refactoredTestSetup.ts';
+import { TestSetup } from '../testSetup.ts';
 import * as utilities from '../utilities/index.ts';
 
 describe('SFDX: Create Project', async () => {
-  const testSetup = new refactoredTestSetup();
+  const testSetup = new TestSetup();
 
   const testReqConfig: utilities.TestReqConfig = {
     projectConfig: {
