+++ test/specs/pushAndPull.e2e.ts
@@ -8,7 +8,7 @@
 import fs from 'fs';
 import { step, xstep } from 'mocha-steps';
 import path from 'path';
-import { refactoredTestSetup } from '../refactoredTestSetup.ts';
+import { TestSetup } from '../testSetup.ts';
 import * as utilities from '../utilities/index.ts';
 import { Workbench } from 'wdio-vscode-service';
 
@@ -32,7 +32,7 @@ describe('Push and Pull', async () => {
   let projectName = '';
   let adminName = '';
   let adminEmailAddress = '';
-  const testSetup = new refactoredTestSetup();
+  const testSetup = new TestSetup();
   const testReqConfig: utilities.TestReqConfig = {
     projectConfig: {
       projectShape: utilities.ProjectShapeOption.NEW,
